Fixed load function (was using parseInt for prices which are double), lowered mill production times, made corn price show on button earlier

+++ LoadAction.java
@@ -25,7 +25,7 @@ public class LoadAction implements Action {
 					// reads in all wheat variables
 					reader.nextLine( );
 					WheatField.owned = Integer.parseInt( reader.nextLine( ) );
-					WheatField.price = Integer.parseInt( reader.nextLine( ) );
+					WheatField.price = Double.parseDouble( reader.nextLine( ) );
 					Framework.wheat = Integer.parseInt( reader.nextLine( ) );
 					WheatField.updateText( );
 
@@ -50,7 +50,7 @@ public class LoadAction implements Action {
 					// reads in all corn variables
 					reader.nextLine( );
 					CornField.owned = Integer.parseInt( reader.nextLine( ) );
-					CornField.price = Integer.parseInt( reader.nextLine( ) );
+					CornField.price = Double.parseDouble( reader.nextLine( ) );
 					Framework.corn = Integer.parseInt( reader.nextLine( ) );
 					CornField.updateText( );
 
@@ -74,7 +74,7 @@ public class LoadAction implements Action {
 				if ( reader.findWithinHorizon( "mill", 99999 ) != null) {
 					reader.nextLine( );
 					Mill.level = Integer.parseInt( reader.nextLine( ) );
-					Mill.price = Integer.parseInt( reader.nextLine( ) );
+					Mill.price = Double.parseDouble( reader.nextLine( ) );
 					Framework.flour = Integer.parseInt( reader.nextLine( ) );
 					Framework.cornMeal = Integer.parseInt( reader.nextLine( ) );
 				}
